# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  order: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  order?: number;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  order?: number;
}

export interface ReorderInput {
  ids: number[];
}

export interface Material {
  id: number;
  categoryId: number;
  name: string;
  unit: string;
  /** @nullable */
  description?: string | null;
  order: number;
  createdAt: string;
}

export interface MaterialInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  description?: string;
  order?: number;
}

export interface MaterialUpdate {
  categoryId?: number;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  unit?: string;
  description?: string;
  order?: number;
}

export interface Connection {
  id: number;
  name: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  copiedFromId?: number | null;
  createdAt: string;
}

export interface ConnectionInput {
  /** @minLength 1 */
  name: string;
  note?: string;
}

export interface ConnectionUpdate {
  /** @minLength 1 */
  name?: string;
  note?: string;
}

export interface ConnectionCopyInput {
  /** @minLength 1 */
  name: string;
  note?: string;
}

export interface ConnectionItem {
  id: number;
  connectionId: number;
  materialId: number;
  quantity: number;
}

export interface ConnectionItemInput {
  materialId: number;
  quantity: number;
}

export interface ConnectionItemsBatch {
  items: ConnectionItemInput[];
}

export interface ConnectionDetail {
  id: number;
  name: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  copiedFromId?: number | null;
  createdAt: string;
  items: ConnectionItem[];
}

export interface SummaryRow {
  materialId: number;
  materialName: string;
  unit: string;
  categoryId: number;
  categoryName: string;
  totalQuantity: number;
}

export interface Summary {
  rows: SummaryRow[];
  connectionCount: number;
}

export type ListMaterialsParams = {
categoryId?: number;
};