# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */

export interface Connection {
  id: number;
  name: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  copiedFromId?: number | null;
  createdAt: Date;
}