# lib/api-zod/src/generated/types/connectionDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionItem } from './connectionItem';

export interface ConnectionDetail {
  id: number;
  name: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  copiedFromId?: number | null;
  createdAt: Date;
  items: ConnectionItem[];
}