# lib/api-zod/src/generated/types/connectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionInput {
  /** @minLength 1 */
  name: string;
  note?: string;
}