# lib/api-zod/src/generated/types/connectionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionItem {
  id: number;
  connectionId: number;
  materialId: number;
  quantity: number;
}