# lib/api-zod/src/generated/types/connectionItemsBatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionItemInput } from './connectionItemInput';

export interface ConnectionItemsBatch {
  items: ConnectionItemInput[];
}