# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Evidence materiálu plynovodních přípojek
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  description?: string;
  order?: number;
}